fix(automation): remove writeRun/writeRunStep/updateRun/updateRunStep actions (closes #3939)
Inline all Supabase writes from the deleted supabase/automationRuns.ts and
supabase/automationRunSteps.ts action files directly into processRun
(internalAction), which already has a createSupabaseDb() instance. The run
insert is moved from emitEvent's scheduled writeRun into the top of processRun;
upsert-on-id semantics handle action retries safely and the UNIQUE constraint
on eventIdempotencyKey still prevents duplicate rows.

For the four updateRunStep call sites inside internalMutations
(_recordAutomationEmailResult, _recordUpdateFieldResult), a new
_patchAutomationRunStep internalAction replaces the deleted external action.
